+++ src/messages.rs
@@ -25,7 +25,7 @@ use sodiumoxide::crypto::{box_, sign};
 use std::fmt::{self, Debug, Formatter};
 
 use authority::Authority;
-use data::{Data, DataRequest};
+use data::{Data, DataIdentifier};
 use error::RoutingError;
 use id::{FullId, PublicId};
 use types::MessageId;
@@ -302,7 +302,7 @@ pub enum RequestContent {
     Refresh(Vec<u8>, MessageId),
     // ---------- External ------------
     /// Ask for data from network, passed from API with data name as parameter
-    Get(DataRequest, MessageId),
+    Get(DataIdentifier, MessageId),
     /// Put data to network. Provide actual data as parameter
     Put(Data, MessageId),
     /// Post data to network. Provide actual data as parameter
@@ -363,11 +363,11 @@ pub enum ResponseContent {
     /// may be shortcut if the data is in a node's cache.
     GetSuccess(Data, MessageId),
     /// Success token for Put (may be ignored)
-    PutSuccess(MessageId),
+    PutSuccess(DataIdentifier, MessageId),
     /// Success token for Post  (may be ignored)
-    PostSuccess(MessageId),
+    PostSuccess(DataIdentifier, MessageId),
     /// Success token for delete  (may be ignored)
-    DeleteSuccess(MessageId),
+    DeleteSuccess(DataIdentifier, MessageId),
     /// Error for `Get`, includes signed request to prevent injection attacks
     GetFailure {
         /// Unique message identifier
@@ -531,14 +531,17 @@ impl Debug for ResponseContent {
             ResponseContent::GetSuccess(ref data, ref message_id) => {
                 write!(formatter, "GetSuccess {{ {:?}, {:?} }}", data, message_id)
             }
-            ResponseContent::PutSuccess(ref message_id) => {
-                write!(formatter, "PutSuccess {{ {:?} }}", message_id)
+            ResponseContent::PutSuccess(ref name, ref message_id) => {
+                write!(formatter, "PutSuccess {{ {:?}, {:?} }}", name, message_id)
             }
-            ResponseContent::PostSuccess(ref message_id) => {
-                write!(formatter, "PostSuccess {{ {:?} }}", message_id)
+            ResponseContent::PostSuccess(ref name, ref message_id) => {
+                write!(formatter, "PostSuccess {{ {:?}, {:?} }}", name, message_id)
             }
-            ResponseContent::DeleteSuccess(ref message_id) => {
-                write!(formatter, "DeleteSuccess {{ {:?} }}", message_id)
+            ResponseContent::DeleteSuccess(ref name, ref message_id) => {
+                write!(formatter,
+                       "DeleteSuccess {{ {:?}, {:?} }}",
+                       name,
+                       message_id)
             }
             ResponseContent::GetFailure { ref id, ref request, .. } => {
                 write!(formatter, "GetFailure {{ {:?}, {:?}, .. }}", id, request)