+++ src/node.rs
@@ -25,7 +25,7 @@ use std::sync::mpsc::{Receiver, Sender, channel};
 use action::Action;
 use authority::Authority;
 use core::Core;
-use data::{Data, DataRequest};
+use data::{Data, DataIdentifier};
 use error::{InterfaceError, RoutingError};
 use event::Event;
 use messages::{RequestContent, RequestMessage, ResponseContent, ResponseMessage, RoutingMessage};
@@ -99,7 +99,7 @@ impl Node {
     }
 
     #[cfg(feature = "use-mock-crust")]
-    #[allow(missing_docs)]
+    /// Poll and process all events in this node's `Core` instance.
     pub fn poll(&self) -> bool {
         self.core.borrow_mut().poll()
     }
@@ -108,7 +108,7 @@ impl Node {
     pub fn send_get_request(&self,
                             src: Authority,
                             dst: Authority,
-                            data_request: DataRequest,
+                            data_request: DataIdentifier,
                             id: MessageId)
                             -> Result<(), InterfaceError> {
         let routing_msg = RoutingMessage::Request(RequestMessage {
@@ -203,12 +203,13 @@ impl Node {
     pub fn send_put_success(&self,
                             src: Authority,
                             dst: Authority,
+                            name: DataIdentifier,
                             id: MessageId)
                             -> Result<(), InterfaceError> {
         let routing_msg = RoutingMessage::Response(ResponseMessage {
             src: src,
             dst: dst,
-            content: ResponseContent::PutSuccess(id),
+            content: ResponseContent::PutSuccess(name, id),
         });
         self.send_action(routing_msg)
     }
@@ -237,12 +238,13 @@ impl Node {
     pub fn send_post_success(&self,
                              src: Authority,
                              dst: Authority,
+                             name: DataIdentifier,
                              id: MessageId)
                              -> Result<(), InterfaceError> {
         let routing_msg = RoutingMessage::Response(ResponseMessage {
             src: src,
             dst: dst,
-            content: ResponseContent::PostSuccess(id),
+            content: ResponseContent::PostSuccess(name, id),
         });
         self.send_action(routing_msg)
     }
@@ -271,12 +273,13 @@ impl Node {
     pub fn send_delete_success(&self,
                                src: Authority,
                                dst: Authority,
+                               name: DataIdentifier,
                                id: MessageId)
                                -> Result<(), InterfaceError> {
         let routing_msg = RoutingMessage::Response(ResponseMessage {
             src: src,
             dst: dst,
-            content: ResponseContent::DeleteSuccess(id),
+            content: ResponseContent::DeleteSuccess(name, id),
         });
         self.send_action(routing_msg)
     }
@@ -418,7 +421,7 @@ impl Drop for Node {
 //         client.put(data.clone());
 //         ::std::thread::sleep_ms(5000);
 
-//         let recovered_data = match client.get(::data::DataRequest::PlainData(name)) {
+//         let recovered_data = match client.get(::data::DataIdentifier::PlainData(name)) {
 //             Some(data) => data,
 //             None => panic!("Failed to recover stored data: {}.", name),
 //         };